Removed all exceptions, since most of them weren't being caught anyway. Added some simple ERROR/WARNING/ASSERT macros that we can use for error-checking during development (or beyond if we want.  All exceptions were replaced with macros.
We should discuss this especially since one test now fails due to relying
on a caught exception.

Note that these macros are located in the datastore for now, but they could
be used more generally.  Also, they will probably be replaced with something
more sophisticated in the future, or they could be enhanced easily. For
example, they only use exit() and print to std out so they don't handle
shutting down an MPI code gracefully.