ENH: initial implementation of the logging api
This commits introduces an initial implementation of
the logging facility. The two key classes added are
Logger and LogStream. Logger is the top level class
that the application uses either directly or through
convenience macros. LogStream is an abstract base
class which represents a stream where the messages
will be logged. The stream can be either a file,
the console (i.e., stdout) etc. It is a pure abstract
class and it defines a unified interface that concrete
implementations must implement. Consequently, it provides
the mechanism by which applications can prescribe a
customized LogStream to fit their needs.

As a pedagogical example and for convenience, this commit
provides two concrete instances of the LogStream: (a)
Console and (b) SynchronizedConsole. Console simply
dumps messages to 'stdout'. SynchronizedConsole dumps
messages to 'stdout', but it is MPI-Aware in that it
dumps the messages in rank order when flushAllStreams()
is called, a collective operation.

Last, two simple examples illustrating the use of the
present implementation are added in examples/logging_example.cc
and examples/parallel_logging_example.cc