ENH: add different variants of logMessage()
This commit adds different variants of logMessage(...), e.g.,
with file,line number without etc. Notably, a tag field was
added to allow applications to associate messages with a given
tag. A tag can be used for example to distinguish messages from
different components and filter them out in post-processing.