ENH: Added ability to timestamp messages
Modified LogStream::getFormattedMessage() to get a textual
representation of the current date and time if the
<TIMESTAMP> key exists in the user-supplied format string,
per Kenny's suggestion. Further, modified the "logging_example.cc"
to use a custom format to illustrate (a) the use of time-stamping
log messages and (b) how an application can prescribe the format
string. Improved the documentation to reflect this changes.